MyYOLOProject/requirements.txt
ultralytics>=8.3.0
pyyaml
torch>=2.0.0
torchvision>=0.15.0
numpy
pandas
matplotlib
seaborn
tqdm

tensorflow>=2.13.0
tf_keras<=2.19.0

onnx>=1.16.0
onnxruntime
onnxsim>=0.4.0
onnx2tf>=1.26.3
onnxslim>=0.1.71
sng4onnx>=1.0.1
onnx_graphsurgeon>=0.3.26
ai-edge-litert>=1.2.0

protobuf>=3.20.0
coremltools>=7.0